Moved the Notifications source file to the public API.
